xatolar tuzaldi

+++ src/main/java/org/example/moreeduceorigin/repository/CommentRepo.java
@@ -3,5 +3,5 @@ package org.example.moreeduceorigin.repository;
 import org.example.moreeduceorigin.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepo extends JpaRepository<Comment, Integer> {
+public interface CommentRepo extends JpaRepository<Comment, Long> {
 }
