+++ src/main/java/org/example/moreeduceorigin/repository/VideoRepo.java
@@ -3,5 +3,5 @@ package org.example.moreeduceorigin.repository;
 import org.example.moreeduceorigin.model.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VideoRepo extends JpaRepository<Video, Integer> {
+public interface VideoRepo extends JpaRepository<Video, Long> {
 }
